+++ test/integration_tests/main.dart
@@ -1,0 +1,10 @@
+// Flutter dependencies
+
+// Project dependencies
+
+void runAllIntegrationTests () {
+
+
+}
+
+