+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### 1.1.2 
+
+- Added: The base structure for API requests.
+
+### 1.1.1 
+
+- Changed: Migrated from [SharedPreferences] to [GetStorage]
+
+## 1.1.0 
+
+- Added: A menu tile animation
+- Added: A content cover animation
+- Updated: The dependencies
+
 # 1.0.0
 
 - Created: The website structure